+++ docs/changelog.md
@@ -6,6 +6,7 @@
      - ⚠️ Please note that the `apprise-api` config is now `apprise_api`
      - Docker secrets are now automatically detected & added to configuration
  - Add [option](https://frigate-notify.0x2142.com/latest/config/file/#general) to collect higher resolution snapshot from camera main stream
+ - Add ability to send video clip instead of snapshot image via [Telegram](https://frigate-notify.0x2142.com/latest/config/file/#telegram)
 
 ## [v0.4.4](https://github.com/0x2142/frigate-notify/releases/tag/v0.4.4) - Apr 08 2025
  - Add support for notifications via [Matrix](https://frigate-notify.0x2142.com/latest/config/file/#matrix)